migratted authenticator functionalities from user service to auth service
